% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Annotate_Cell_Community}
\alias{Annotate_Cell_Community}
\title{Function to annotate each cell with corresponding spatial community
and spatial neighborhood, and return final metadata}
\usage{
Annotate_Cell_Community(
  cell_spatial_community_cluster,
  spatial_community_anno,
  metadata
)
}
\description{
Function to annotate each cell with corresponding spatial community
and spatial neighborhood, and return final metadata
}
