% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Create_Montage_Heatmap}
\alias{Create_Montage_Heatmap}
\title{Function to create a heatmap for the montage displaying
the spatial community changes along the regions}
\usage{
Create_Montage_Heatmap(metadata, gene_set_name)
}
\description{
Function to create a heatmap for the montage displaying
the spatial community changes along the regions
}
