% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Create_Signature_Matrix}
\alias{Create_Signature_Matrix}
\title{The second part hosts the function for creating
a MONTAGE signature matrix (Matrix M)
nrow=number of marker genes,ncol=number of communities
from single-cell RNA sequencing (Matrix G)
and spatial community cell type compositions (Matrix S)}
\usage{
Create_Signature_Matrix(Matrix_G_file, Matrix_S_file)
}
\description{
The second part hosts the function for creating
a MONTAGE signature matrix (Matrix M)
nrow=number of marker genes,ncol=number of communities
from single-cell RNA sequencing (Matrix G)
and spatial community cell type compositions (Matrix S)
}
