% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Enrichment_Calculation}
\alias{Enrichment_Calculation}
\title{The third part hosts the functions for identifying spatial community
composition changes (aka:montage) along a certain functional enrichment
in tissue
Function to calculate gene enrichment scores.Input to this function:
(1) The MONTAGE spatial community gene signature matrix
(2) Gene set for biological functions, such as Hallmark genesets from MsigDB
The geneset input needs to be a dataframe with each row containing gene
signatures for a biological function}
\usage{
Enrichment_Calculation(gene_sets, SC_signature)
}
\description{
The third part hosts the functions for identifying spatial community
composition changes (aka:montage) along a certain functional enrichment
in tissue
Function to calculate gene enrichment scores.Input to this function:
(1) The MONTAGE spatial community gene signature matrix
(2) Gene set for biological functions, such as Hallmark genesets from MsigDB
The geneset input needs to be a dataframe with each row containing gene
signatures for a biological function
}
