% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Identify_Gradient}
\alias{Identify_Gradient}
\title{Function to identify the regions along a spatial gradient
of a certain biological functional enrichment}
\usage{
Identify_Gradient(metadata, gene_set_name, enrichment_scores, samples)
}
\description{
Function to identify the regions along a spatial gradient
of a certain biological functional enrichment
}
