% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{KNN_Neighbors}
\alias{KNN_Neighbors}
\title{The first part hosts the functions for spatial community identification
in single-cell spatial data
Function to find N-nearest neighboring cells}
\usage{
KNN_Neighbors(coords, number_of_neighbors)
}
\description{
The first part hosts the functions for spatial community identification
in single-cell spatial data
Function to find N-nearest neighboring cells
}
