% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Neighborhood_Density}
\alias{Neighborhood_Density}
\title{Function to calculate cell type densities for each cell neighborhood}
\usage{
Neighborhood_Density(metadata, number_of_neighbors = 10, cell_types, sample_id)
}
\description{
Function to calculate cell type densities for each cell neighborhood
}
