% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Cells}
\alias{Plot_Cells}
\title{Function to plot cell types that can be mapped to original images
Each cell is represented by a dot indicating the centroid with XY}
\usage{
Plot_Cells(
  metadata,
  sample_to_check,
  cell_types_to_plot,
  cell_type_colors = NULL,
  test_size = 1
)
}
\description{
Function to plot cell types that can be mapped to original images
Each cell is represented by a dot indicating the centroid with XY
}
