% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Community}
\alias{Plot_Community}
\title{Function to plot cell spatial community mapped to original images}
\usage{
Plot_Community(
  metadata,
  sample_to_check,
  community_to_plot,
  community_colors = NULL,
  test_size
)
}
\description{
Function to plot cell spatial community mapped to original images
}
