% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Community_in_Region}
\alias{Plot_Community_in_Region}
\title{Function to plot cell spatial communities for a certain region
along the spatial gradient}
\usage{
Plot_Community_in_Region(
  metadata,
  sample_to_check,
  community_to_check,
  region_to_plot,
  community_colors = NULL,
  test_size = 0.1
)
}
\description{
Function to plot cell spatial communities for a certain region
along the spatial gradient
}
