% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Enrichment_Voronoi}
\alias{Plot_Enrichment_Voronoi}
\title{Functional to plot the enrichment scores back onto the single cells in the image}
\usage{
Plot_Enrichment_Voronoi(
  metadata,
  sample_to_check,
  enrichment_scores,
  gene_set_name,
  test_size = 0.1
)
}
\description{
Functional to plot the enrichment scores back onto the single cells in the image
}
