% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Neighborhood}
\alias{Plot_Neighborhood}
\title{Function to plot cell spatial neighborhood mapped to original images}
\usage{
Plot_Neighborhood(
  metadata,
  sample_to_check,
  neighborhood_to_plot,
  neighborhood_colors = NULL,
  test_size = 1
)
}
\description{
Function to plot cell spatial neighborhood mapped to original images
}
