% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Plot_Regions}
\alias{Plot_Regions}
\title{Function to plot regions along spatial gradient for each cell in the image}
\usage{
Plot_Regions(metadata, sample_to_plot, gene_set_name, test_size = 0.1)
}
\description{
Function to plot regions along spatial gradient for each cell in the image
}
