% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Spatial_Community_Clustering}
\alias{Spatial_Community_Clustering}
\title{Function to cluster the spatial neighborhood cell compositions
and obtain spatial community cell compositions}
\usage{
Spatial_Community_Clustering(
  initial_neighborhood_clustering,
  neighborhood_cell_composition,
  metadata,
  n_community = 8
)
}
\description{
Function to cluster the spatial neighborhood cell compositions
and obtain spatial community cell compositions
}
