% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MONTAGE_R_functions.R
\name{Spatial_Neighborhood_Clustering}
\alias{Spatial_Neighborhood_Clustering}
\title{Function to obtain the neighborhood for each cell after clustering}
\usage{
Spatial_Neighborhood_Clustering(
  n_neighborhood = 30,
  neighbor_matrix,
  metadata,
  cell_types,
  cluster_center = NULL
)
}
\description{
Function to obtain the neighborhood for each cell after clustering
}
